Update handler path
Changes disassociation endpoint to `/api/unlink.json`.
Moves handler to `plugins/openlibrary/api.py`.
